Improve similarity handling and update permissions
Updates the similarity handling in the vector search strategy to directly use the similarity value instead of converting it to a double.

Additionally, adds 'target/' to the .gitignore file to prevent compilation artifacts from being tracked, and modifies the init-es.sh script to make it executable.

+++ src/main/java/com/enterprise/rag/domain/service/search/VectorSearchStrategy.java
@@ -73,7 +73,7 @@ public class VectorSearchStrategy implements SearchStrategy {
                 .numCandidates(numCandidates);
         
         if (similarity != null) {
-            knnBuilder.similarity(similarity.doubleValue());
+            knnBuilder.similarity(similarity);
         }
         
         return knnBuilder.build();
